refactor: [MUS-14] reorganizacion de carpetas. aplicacion de metodologia Screaming
