+++ src/app/page.tsx
@@ -1,12 +1,11 @@
-"use client"
-import Image from 'next/image';
-import { useAuth } from '@/app/Store/authStore';
-import LoginForm from '@/components/LoginForm';
-import ToggleSwitch from '@/components/UI/Toggle';
+"use client";
+import Image from "next/image";
+import { useAuth } from "@/domains/auth/Store/authStore";
+import LoginForm from "@/domains/auth/LoginForm";
+import ToggleSwitch from "@/shared/components/UI/Toggle";
 
 export default function LoginPage() {
   const { isLoggedIn, login } = useAuth();
-
 
   const handleLoginSuccess = (email: string) => {
     login(email);
@@ -22,8 +21,17 @@ export default function LoginPage() {
         <ToggleSwitch />
       </div>
       <div className="flex flex-col items-center justify-center px-6 py-8 mx-auto md:h-screen lg:py-0">
-        <a href="#" className="flex items-center mb-6 text-2xl font-semibold text-text-main dark:text-gray-100">
-          <Image className="w-12 h-10 mr-2 rounded-lg" src="/logo.webp" alt="logo" width={36} height={36} />
+        <a
+          href="#"
+          className="flex items-center mb-6 text-2xl font-semibold text-text-main dark:text-gray-100"
+        >
+          <Image
+            className="w-12 h-10 mr-2 rounded-lg"
+            src="/logo.webp"
+            alt="logo"
+            width={36}
+            height={36}
+          />
           Musila
         </a>
         <div className="w-full bg-background-light rounded-lg shadow-lg dark:border md:mt-0 sm:max-w-md xl:p-0 dark:bg-gray-800 dark:border-gray-700">
