--- src/domains/tracks/services/graphql/querys/queryTest.ts
@@ -1,9 +1,0 @@
-import { gql } from "@apollo/client";
-
-//TODO: Copiar la query
-export const TRACK_QUERY = gql`
-  query {
-    
-  }
-
-`;
