--- src/domains/tracks/types/tracks.ts
@@ -1,7 +1,0 @@
-export interface Track {
-  id: string;
-  name: string;
-  url: string;
-  cover: string;
-  //demas propiedades...etc
-}
