+++ src/shared/components/Layout/ContentApp.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import React from 'react';
-import ToggleSwitch from '@/components/UI/Toggle';
+import React from "react";
+import ToggleSwitch from "@/shared/components/UI/Toggle";
 
 export default function ContentApp({
   children,
-  onToggleSidebar, 
+  onToggleSidebar,
 }: Readonly<{
   children: React.ReactNode;
   onToggleSidebar: () => void;
@@ -20,19 +20,35 @@ export default function ContentApp({
           aria-label="Abrir menú"
         >
           {/* Icono de hamburguesa */}
-          <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M4 6h16M4 12h16M4 18h16"></path></svg>
+          <svg
+            className="w-6 h-6"
+            fill="none"
+            stroke="currentColor"
+            viewBox="0 0 24 24"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              d="M4 6h16M4 12h16M4 18h16"
+            ></path>
+          </svg>
         </button>
-        <span className="text-xl font-semibold text-text-main dark:text-white">Musila App</span>
+        <span className="text-xl font-semibold text-text-main dark:text-white">
+          Musila App
+        </span>
         <ToggleSwitch />
       </header>
 
-
       <div className="flex-1 p-6 bg-white dark:bg-gray-900">
         <div className="flex items-center justify-between mb-6">
-          <h1 className="text-2xl font-bold text-text-main dark:text-white">Contenido de la Aplicación</h1>
+          <h1 className="text-2xl font-bold text-text-main dark:text-white">
+            Contenido de la Aplicación
+          </h1>
 
           <div className="hidden md:block">
-              <ToggleSwitch />
+            <ToggleSwitch />
           </div>
         </div>
         {children}
